requirements.txt
numpy
pandas
scikit-learn
catboost
xgboost
dill
flask
flask-cors
python-dotenv
pyodbc
sqlalchemy
requests
openpyxl
selenium
geopandas
uszipcode
